Simplify preload hook
