+++ src/preload/hook-ipc.ts
@@ -2,9 +2,6 @@ import { contextBridge, ipcRenderer } from 'electron';
 
 import { IPC } from '../ipc';
 
-/** Timeout used for registering with the main process. */
-const PING_TIMEOUT = 500;
-
 const ipcObject = {
   sendScope: (scopeJson: string) => ipcRenderer.send(IPC.SCOPE, scopeJson),
   sendEvent: (eventJson: string) => ipcRenderer.send(IPC.EVENT, eventJson),
@@ -20,20 +17,10 @@ window.__SENTRY_IPC__ = ipcObject;
 
 // We attempt to use contextBridge if it's available (Electron >= 6)
 if (contextBridge) {
-  // This will fail if contextIsolation is not enabled but we have no other way to detect this from the renderer
+  // This will fail if contextIsolation is not enabled but we have no other way to detect this from the preload
   try {
     contextBridge.exposeInMainWorld('__SENTRY_IPC__', ipcObject);
   } catch (e) {
     //
   }
 }
-
-// Checks if the main processes is available and logs a warning if not.
-setTimeout(() => {
-  const timeout = setTimeout(() => {
-    // eslint-disable-next-line no-console
-    console.warn('Could not connect to Sentry main process. Did you call init in the Electron main process?');
-  }, PING_TIMEOUT);
-
-  window.__SENTRY_IPC__?.pingMain(() => clearTimeout(timeout));
-}, PING_TIMEOUT);
